클리어 만들기위한단계1 - 기본벽돌 모두 소진시 맵이동 완료

+++ Blitz_API/EngineBase/EngineRandom.h
@@ -17,7 +17,7 @@ public:
 	{
 		if (_Max < _Min)
 		{
-			float SwapValue = _Min;
+			int SwapValue = _Min;
 			_Min = _Max;
 			_Max = SwapValue;
 		}
